Moved components into view folder
