+++ src/view/components/__tests__/LocationDropdown.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import { CountryDropdown } from "components/CountryDropdown";
 import "@testing-library/jest-dom";
+import { CountryDropdown } from "../country-dropdown/CountryDropdown";
 
 describe("CountryDropdown", () => {
     it("Dropdown button should show add a country when no countries are selected", () => {
