+++ src/view/components/country-dropdown/CountryDropdown.tsx
@@ -4,7 +4,7 @@ import { PlusIcon } from "@heroicons/react/20/solid";
 import {
     COUNTRY_CO2_EMISSIONS_GRAMS_PER_GB,
     CountryName,
-} from "../data/constants/CountryEmissions";
+} from "../../../data/constants/CountryEmissions";
 
 type CountryDropdownType = {
     addSelectedCountry: (country: CountryName) => void;
@@ -33,23 +33,23 @@ export const CountryDropdown = ({
                         leaveTo="opacity-0"
                     >
                         <Listbox.Options className="absolute mt-1 max-h-60 w-full overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
-                            {Object.entries(COUNTRY_CO2_EMISSIONS_GRAMS_PER_GB).map(
-                                ([country]) => (
-                                    <Listbox.Option
-                                        key={country}
-                                        className={({ active }) =>
-                                            `relative cursor-default select-none py-2 pl-10 pr-4 ${
-                                                active
-                                                    ? "bg-amber-100 text-amber-900"
-                                                    : "text-gray-900"
-                                            }`
-                                        }
-                                        value={country}
-                                    >
-                                        {country}
-                                    </Listbox.Option>
-                                )
-                            )}
+                            {Object.entries(
+                                COUNTRY_CO2_EMISSIONS_GRAMS_PER_GB
+                            ).map(([country]) => (
+                                <Listbox.Option
+                                    key={country}
+                                    className={({ active }) =>
+                                        `relative cursor-default select-none py-2 pl-10 pr-4 ${
+                                            active
+                                                ? "bg-amber-100 text-amber-900"
+                                                : "text-gray-900"
+                                        }`
+                                    }
+                                    value={country}
+                                >
+                                    {country}
+                                </Listbox.Option>
+                            ))}
                         </Listbox.Options>
                     </Transition>
                 </div>
